replaced float with int

+++ percentagecalc.py
@@ -1,8 +1,8 @@
-n1=float(input("enter marks in english:"))
-n2=float(input("enter marks in hindi:"))
-n3=float(input("enter marks in maths:"))
-n4=float(input("enter marks in science:"))
-n5=float(input("enter marks in sst:"))
+n1= int(input("enter marks in english:"))
+n2= int(input("enter marks in hindi:"))
+n3= int(input("enter marks in maths:"))
+n4= int(input("enter marks in science:"))
+n5= int(input("enter marks in sst:"))
 per=(n1+n2+n3+n4+n5)/5
 if per >=85:
     print("A grade")
